Se realizaron las validaciones de de volumenes, se agrego la clase abEstadistica by Rkoch3196

+++ src/CalculadoraAvanzada/ManejoErrores.java
@@ -81,11 +81,11 @@ public class ManejoErrores
         return valor;
     }
     
-    public double intValidarExcepciones (String mensaje) throws NumberFormatException 
+    public int intValidarExcepciones (String mensaje) throws NumberFormatException 
     {
         boolean todoBien = true; 
         
-        double valor = 0;
+        int valor = 0;
         
         do
         {
